New examples for the SDL and NanoVG renderers are in /examples - As a part of this big change, the test environment code has been completely removed from the library. - This also means the demo project was replaced and removed. - The xwfont.h main header will now allow the use of multiple renderer backends. - Fixed a type size warning in basic_renderer. - The nanovg renderer needs it's own OpenGL loader (prefixed modified glad source). - I may forget in the future, but the OpenGL loader needs to get called for shared library builds of the library!

+++ src/basic_renderer.cpp
@@ -49,7 +49,7 @@ basic_renderer::basic_renderer(uint32_t * source_pixels, const uint32_t width, c
 basic_renderer::basic_renderer(const FT_Bitmap* ftbitmap, const uint32_t foreground, const uint32_t background)
     : basic_renderer(ftbitmap->width, ftbitmap->rows)
 {
-    for (int32_t i = 0; i < size(); ++i)
+    for (size_t i = 0; i < size(); ++i)
         m_data[i] = blend_colors(
             pixel_rgba(foreground, ftbitmap->buffer[i]),
             background
